# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TokenInsights REST API
  version: 2.0.0
  description:
    Query dataset-scoped usage and ingest sanitized collector evidence. Personal servers are
    unauthenticated; hosted servers authenticate bearer tokens or read-only browser sessions.
    Servers interpret evidence and never scan host artifacts.
...
paths:
  /api/v2/collector-progress:
    get:
      security:
        - bearerAuth: []
        - browserSession: []
        - {}
      operationId: getCollectorProgress
      summary: Read ephemeral personal collector progress
      description:
        Available only when collector-progress capability is advertised. Never starts collection.
        Hosted servers do not expose this route.
      tags:
        - Sync
      responses:
        '200':
          description: Recent local collector attempts, with bounded retention.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectorProgressResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/v2/instance:
    get:
      security:
        - bearerAuth: []
        - browserSession: []
        - {}
      operationId: getInstanceV2
      summary: Describe this TokenInsights server
      tags:
        - Instance
      responses:
        '200':
          description: Server versions, saved data hostname, capabilities, and viewer defaults
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstanceResponseV2'
        default:
          $ref: '#/components/responses/Error'
  /api/v2/status:
    get:
      security:
        - bearerAuth: []
        - browserSession: []
        - {}
      operationId: getStatusV2
      summary: Read processing readiness and committed revision
      tags:
        - Sync
      responses:
        '200':
          description: Saved processing state; never starts collection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponseV2'
        default:
          $ref: '#/components/responses/Error'
  /api/v2/usage:
    get:
      security:
        - bearerAuth: []
        - browserSession: []
        - {}
      operationId: getUsageV2
      summary: Query canonical token usage
      description: 'Returns a sorted, paginated table, complete filtered summary, and chart

        from one validated read transaction. Date bounds use server-local dates.

        '
      tags:
        - Usage
      parameters:
        - $ref: '#/components/parameters/PeriodFilter'
        - $ref: '#/components/parameters/BucketFilter'
        - $ref: '#/components/parameters/FromFilter'
        - $ref: '#/components/parameters/ToFilter'
        - $ref: '#/components/parameters/ProviderFilter'
        - $ref: '#/components/parameters/ModelFilter'
        - $ref: '#/components/parameters/HarnessFilter'
        - $ref: '#/components/parameters/SessionFilter'
        - $ref: '#/components/parameters/RepositoryFilter'
        - $ref: '#/components/parameters/DirectoryFilter'
        - name: tab
          in: query
          description: Aggregation shown in table and chart rows.
          schema:
            allOf:
              - $ref: '#/components/schemas/UsageTab'
            default: tokens
        - name: locationGroup
          in: query
          description: Repo-tab grouping by repository or directory.
          schema:
            $ref: '#/components/schemas/LocationGroup'
        - name: sort
          in: query
          description: Sort field. Context-only fields are valid only for the context tab.
          schema:
            $ref: '#/components/schemas/SortField'
        - name: direction
          in: query
          schema:
            $ref: '#/components/schemas/SortDirection'
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - $ref: '#/components/parameters/QualityFilter'
      responses:
        '200':
          description: Canonical usage result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageResponseV2'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '503':
          $ref: '#/components/responses/Unavailable'
        default:
          $ref: '#/components/responses/Error'
  /api/v2/usage/facets:
    get:
      security:
        - bearerAuth: []
        - browserSession: []
        - {}
      operationId: getUsageFacetsV2
      summary: Query available usage filters
      description: 'Each facet applies the other active filters while omitting its own.

        Session search is a case-insensitive literal substring search capped at

        100 results.

        '
      tags:
        - Usage
      parameters:
        - $ref: '#/components/parameters/PeriodFilter'
        - $ref: '#/components/parameters/BucketFilter'
        - $ref: '#/components/parameters/FromFilter'
        - $ref: '#/components/parameters/ToFilter'
        - $ref: '#/components/parameters/ProviderFilter'
        - $ref: '#/components/parameters/ModelFilter'
        - $ref: '#/components/parameters/HarnessFilter'
        - $ref: '#/components/parameters/SessionFilter'
        - $ref: '#/components/parameters/RepositoryFilter'
        - $ref: '#/components/parameters/DirectoryFilter'
        - name: search
          in: query
          description: Literal substring used to filter session IDs.
          schema:
            type: string
            default: ''
        - name: tab
          in: query
          description: Active aggregation. Repository/directory filters require the repo tab.
          schema:
            allOf:
              - $ref: '#/components/schemas/UsageTab'
            default: tokens
        - $ref: '#/components/parameters/QualityFilter'
      responses:
        '200':
          description: Available dimension values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageFacetsResponseV2'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '503':
          $ref: '#/components/responses/Unavailable'
        default:
          $ref: '#/components/responses/Error'
...
  /api/v2/auth/session:
    post:
      operationId: createBrowserSession
      summary: Exchange bearer token for a read-only browser cookie
      description: Hosted-only, canonical same-origin exchange. Sets a host-only Secure HttpOnly SameSite=Lax cookie for 24 hours. The bearer token is never returned or persisted by the browser.
      tags:
        - Instance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrowserSessionRequest'
      responses:
        '204':
          description: Authenticated browser session established.
        '401':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
    delete:
      operationId: deleteBrowserSession
      summary: End this browser session
      tags:
        - Instance
      responses:
        '204':
          description: Browser session cleared.
        default:
          $ref: '#/components/responses/Error'
  /api/v1/instance:
    get:
      operationId: getInstance
      summary: Describe this TokenInsights server
      tags:
        - Instance
      responses:
        '200':
          description: Server versions, saved data hostname, capabilities, and viewer defaults
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstanceResponse'
              example:
                apiVersion: v1
                serverVersion: 1.0.0
                hostname: workstation
                timezone: Australia/Sydney
                capabilities:
                  - usage
                  - facets
                  - ingestion
                defaults:
                  period: month
                  bucket: day
                  from: ''
                  to: ''
                  providers: []
                  models: []
                  harnesses: []
                  sessions: []
        default:
          $ref: '#/components/responses/Error'
...
  /api/v1/usage:
    get:
      operationId: getUsage
      summary: Query canonical token usage
      description: 'Returns a sorted, paginated table, complete filtered summary, and chart

        from one validated read transaction. Date bounds use server-local dates.

        '
      tags:
        - Usage
      parameters:
        - $ref: '#/components/parameters/PeriodFilter'
        - $ref: '#/components/parameters/BucketFilter'
        - $ref: '#/components/parameters/FromFilter'
        - $ref: '#/components/parameters/ToFilter'
        - $ref: '#/components/parameters/ProviderFilter'
        - $ref: '#/components/parameters/ModelFilter'
        - $ref: '#/components/parameters/HarnessFilter'
        - $ref: '#/components/parameters/SessionFilter'
        - $ref: '#/components/parameters/RepositoryFilter'
        - $ref: '#/components/parameters/DirectoryFilter'
        - name: tab
          in: query
          description: Aggregation shown in table and chart rows.
          schema:
            allOf:
              - $ref: '#/components/schemas/UsageTab'
            default: tokens
        - name: locationGroup
          in: query
          description: Repo-tab grouping by repository or directory.
          schema:
            $ref: '#/components/schemas/LocationGroup'
        - name: sort
          in: query
          description: Sort field. Context-only fields are valid only for the context tab.
          schema:
            $ref: '#/components/schemas/SortField'
        - name: direction
          in: query
          schema:
            $ref: '#/components/schemas/SortDirection'
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - $ref: '#/components/parameters/QualityFilter'
      responses:
        '200':
          description: Canonical usage result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '503':
          $ref: '#/components/responses/Unavailable'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/usage/facets:
    get:
      operationId: getUsageFacets
      summary: Query available usage filters
      description: 'Each facet applies the other active filters while omitting its own.

        Session search is a case-insensitive literal substring search capped at

        100 results.

        '
      tags:
        - Usage
      parameters:
        - $ref: '#/components/parameters/PeriodFilter'
        - $ref: '#/components/parameters/BucketFilter'
        - $ref: '#/components/parameters/FromFilter'
        - $ref: '#/components/parameters/ToFilter'
        - $ref: '#/components/parameters/ProviderFilter'
        - $ref: '#/components/parameters/ModelFilter'
        - $ref: '#/components/parameters/HarnessFilter'
        - $ref: '#/components/parameters/SessionFilter'
        - $ref: '#/components/parameters/RepositoryFilter'
        - $ref: '#/components/parameters/DirectoryFilter'
        - name: search
          in: query
          description: Literal substring used to filter session IDs.
          schema:
            type: string
            default: ''
        - name: tab
          in: query
          description: Active aggregation. Repository/directory filters require the repo tab.
          schema:
            allOf:
              - $ref: '#/components/schemas/UsageTab'
            default: tokens
        - $ref: '#/components/parameters/QualityFilter'
      responses:
        '200':
          description: Available dimension values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageFacetsResponse'
              example:
                providers:
                  - openai
                models:
                  - gpt-5
                harnesses:
                  - codex
                sessions:
                  - session-123
                repositories: []
                directories: []
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '503':
          $ref: '#/components/responses/Unavailable'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Hosted token; effective dataset and read/ingest scopes resolved by
        the server.
    browserSession:
      type: apiKey
      in: cookie
      name: tokeninsights_session
      description: Read-only, HttpOnly browser session established by POST /api/v2/auth/session.
  parameters:
    PeriodFilter:
      name: period
      in: query
      schema:
        allOf:
          - $ref: '#/components/schemas/Period'
        default: month
    BucketFilter:
      name: bucket
      in: query
      schema:
        allOf:
          - $ref: '#/components/schemas/Bucket'
        default: day
    FromFilter:
      name: from
      in: query
      description: Inclusive lower local-date bound; replaces period when supplied.
      schema:
        type: string
        format: date
        x-go-type: string
    ToFilter:
      name: to
      in: query
      description: Inclusive upper local-date bound; replaces period when supplied.
      schema:
        type: string
        format: date
        x-go-type: string
    ProviderFilter:
      name: provider
      in: query
      description: Provider filter. Repeat the parameter to select multiple values.
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
    ModelFilter:
      name: model
      in: query
      description: Model filter. Repeat the parameter to select multiple values.
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
    HarnessFilter:
      name: harness
      in: query
      description: Harness filter. Repeat the parameter to select multiple values.
      style: form
      explode: true
      schema:
        type: array
        items:
          $ref: '#/components/schemas/Harness'
    SessionFilter:
      name: session
      in: query
      description: Session-ID filter. Repeat the parameter to select multiple values.
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
    RepositoryFilter:
      name: repository
      in: query
      description:
        Repo-tab stable repository key; repeat for multiple values. The key unknown selects
        missing values.
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
    DirectoryFilter:
      name: directory
      in: query
      description:
        Repo-tab stable directory key; repeat for multiple values. The key unknown selects
        missing values.
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
    QualityFilter:
      name: quality
      in: query
      schema:
        allOf:
          - $ref: '#/components/schemas/UsageQuality'
        default: confirmed
  responses:
...
    Error:
      description: Request failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    UsageTab:
      type: string
      enum:
        - tokens
        - models
        - providers
        - harnesses
        - sessions
        - context
        - repo
    LocationGroup:
      type: string
      enum:
        - repository
        - directory
    SortField:
      type: string
      enum:
        - name
        - date
        - total
        - input
        - output
        - reasoning
        - cacheRead
        - cacheWrite
        - sessions
        - context
        - averageContext
        - medianContext
        - maxContext
        - harness
        - provider
        - model
    SortDirection:
      type: string
      enum:
        - asc
        - desc
...
    Count:
      type: integer
      format: int64
      minimum: 0
      maximum: 9007199254740991
...
    InstanceResponseV2:
      type: object
      additionalProperties: false
      required:
        - instanceId
        - dataEpoch
        - dataReadiness
        - apiVersion
        - serverVersion
        - hostname
        - timezone
        - capabilities
        - defaults
        - datasetId
        - serverKind
        - permissions
      properties:
        instanceId:
          type: string
          minLength: 1
        dataEpoch:
          type: string
          description: Durable database identity; nonempty when dataReadiness is ready.
        dataReadiness:
          type: string
          enum:
            - ready
            - metadata
            - recovery
            - rebuild
            - unavailable
        apiVersion:
          type: string
          enum:
            - v2
        serverVersion:
          type: string
          minLength: 1
        hostname:
          description:
            'Producer hostname: unknown before ingestion; multiple machines when distinct producer
            labels are present.'
          type: string
          minLength: 1
        timezone:
          type: string
          minLength: 1
          description:
            Server reporting timezone, preferably an IANA name with historical DST rules. Unnamed
            system zones use an explicit UTC offset fallback.
        capabilities:
          type: array
          uniqueItems: true
          items:
            type: string
            minLength: 1
            description:
              'Unknown capabilities are ignored by clients. Known names: usage, facets, web-dashboard,
              raw-ingestion, terminal-dashboard, collector-progress, reprocess.'
        defaults:
          $ref: '#/components/schemas/Selection'
        datasetId:
          type: string
          minLength: 1
          maxLength: 256
        serverKind:
          type: string
          enum:
            - personal
            - hosted
        permissions:
          type: array
          uniqueItems: true
          items:
            type: string
            enum:
              - read
              - ingest
    StatusResponseV2:
      type: object
      additionalProperties: false
      required:
        - instanceId
        - dataEpoch
        - dataReadiness
        - revision
        - datasetId
        - generation
        - targetGeneration
        - inputRevision
        - pending
      properties:
        instanceId:
          type: string
          minLength: 1
        dataEpoch:
          type: string
          description: Durable database identity; nonempty when dataReadiness is ready.
        dataReadiness:
          type: string
          enum:
            - ready
            - metadata
            - recovery
            - rebuild
            - unavailable
        revision:
          $ref: '#/components/schemas/Count'
        generation:
          $ref: '#/components/schemas/Count'
        inputRevision:
          $ref: '#/components/schemas/Count'
        pending:
          $ref: '#/components/schemas/Count'
        datasetId:
          type: string
          minLength: 1
          maxLength: 256
        targetGeneration:
          $ref: '#/components/schemas/Count'
    UsageResponseV2:
      type: object
      additionalProperties: false
      required:
        - instanceId
        - dataEpoch
        - revision
        - rows
        - chart
        - rowCount
        - page
        - pageSize
        - summary
        - lastSynced
        - range
        - datasetId
      properties:
        factCount:
          $ref: '#/components/schemas/Count'
        instanceId:
          type: string
          minLength: 1
        dataEpoch:
          type: string
          minLength: 1
        revision:
          $ref: '#/components/schemas/Count'
        rows:
          type: array
          items:
            $ref: '#/components/schemas/UsageRow'
        chart:
          type: array
          items:
            $ref: '#/components/schemas/UsageRow'
        rowCount:
          $ref: '#/components/schemas/Count'
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 200
        summary:
          $ref: '#/components/schemas/UsageSummary'
        lastSynced:
          description: Time of the last committed ingestion, in Unix milliseconds; zero for an empty server.
          allOf:
            - $ref: '#/components/schemas/Count'
        range:
          type: string
        generation:
          $ref: '#/components/schemas/Count'
        inputRevision:
          $ref: '#/components/schemas/Count'
        pending:
          $ref: '#/components/schemas/Count'
        unresolved:
          $ref: '#/components/schemas/Count'
        quality:
          $ref: '#/components/schemas/UsageQuality'
        datasetId:
          type: string
          minLength: 1
          maxLength: 256
...
    CollectorProgressResponse:
      type: object
      additionalProperties: false
      required:
        - instanceId
        - attempts
      properties:
        instanceId:
          type: string
          minLength: 1
        attempts:
          type: array
          maxItems: 32
          items:
            $ref: '#/components/schemas/CollectorAttempt'